refactor: naming updates
