+++ src/features/listings/components/Listings.tsx
@@ -1,0 +1,188 @@
+import { usePrivy } from '@privy-io/react-auth';
+import { ArrowRight } from 'lucide-react';
+import Link from 'next/link';
+import { usePostHog } from 'posthog-js/react';
+
+import { EmptySection } from '@/components/shared/EmptySection';
+import { Button } from '@/components/ui/button';
+import { Separator } from '@/components/ui/separator';
+import { useBreakpoint } from '@/hooks/use-breakpoint';
+
+import { CATEGORY_NAV_ITEMS } from '@/features/navbar/constants';
+
+import { type ListingCategory, useListings } from '../hooks/useListings';
+import { useListingState } from '../hooks/useListingState';
+import type { ListingTabsProps } from '../types';
+import { CategoryPill } from './CategoryPill';
+import { ListingCard, ListingCardSkeleton } from './ListingCard';
+import { ListingFilters } from './ListingFilters';
+import { ListingTabs } from './ListingTabs';
+
+export const Listings = ({
+  type,
+  potentialSession,
+  region,
+  sponsor,
+}: ListingTabsProps) => {
+  const posthog = usePostHog();
+  const isMd = useBreakpoint('md');
+
+  const { authenticated } = usePrivy();
+
+  const {
+    activeTab,
+    activeCategory,
+    activeStatus,
+    activeSortBy,
+    activeOrder,
+    handleTabChange,
+    handleCategoryChange,
+    handleStatusChange,
+    handleSortChange,
+  } = useListingState({
+    defaultCategory:
+      (potentialSession || authenticated) && type === 'home'
+        ? 'For You'
+        : 'All',
+  });
+
+  const {
+    data: listings,
+    isLoading,
+    error,
+  } = useListings({
+    context: type,
+    tab: activeTab,
+    category: activeCategory,
+    status: activeStatus,
+    sortBy: activeSortBy,
+    order: activeOrder,
+    region,
+    sponsor,
+  });
+
+  const viewAllLink = () => {
+    if (type === 'home') {
+      return '/all';
+    }
+    if (type === 'region') {
+      return `/regions/${region}/all`;
+    } else return '/all';
+  };
+
+  const renderContent = () => {
+    if (isLoading) {
+      return Array.from({ length: 5 }).map((_, index) => (
+        <ListingCardSkeleton key={index} />
+      ));
+    }
+
+    if (error) {
+      return <EmptySection title="Error loading listings" />;
+    }
+
+    if (!listings?.length) {
+      return <EmptySection title="No listings found" />;
+    }
+
+    return (
+      <>
+        {listings.map((listing) => (
+          <ListingCard key={listing.id} bounty={listing} />
+        ))}
+        {(type === 'home' || type === 'region') && (
+          <Button
+            className="my-8 w-full border-slate-300 py-5 text-slate-400"
+            onClick={() => posthog.capture('viewall bottom_listings')}
+            size="sm"
+            variant="outline"
+            asChild
+          >
+            <Link className="ph-no-capture" href={viewAllLink()}>
+              View All
+              <ArrowRight className="ml-2 h-4 w-4" />
+            </Link>
+          </Button>
+        )}
+      </>
+    );
+  };
+
+  return (
+    <div className="mt-5 mb-10">
+      <div className="flex w-full items-center justify-between md:mb-1.5">
+        <div className="flex items-center">
+          <p className="text-lg font-semibold text-slate-800">
+            Browse Opportunities
+          </p>
+
+          <div className="hidden items-center md:flex">
+            <Separator orientation="vertical" className="mx-3 h-6" />
+            <ListingTabs
+              activeTab={activeTab}
+              handleTabChange={handleTabChange}
+            />
+          </div>
+        </div>
+
+        <ListingFilters
+          activeStatus={activeStatus}
+          activeSortBy={activeSortBy}
+          activeOrder={activeOrder}
+          onStatusChange={handleStatusChange}
+          onSortChange={handleSortChange}
+        />
+      </div>
+      <div className="mt-2 mb-1 md:hidden">
+        <ListingTabs activeTab={activeTab} handleTabChange={handleTabChange} />
+      </div>
+
+      <div className="mb-2 h-px w-full bg-slate-200" />
+
+      <div className="flex gap-1 overflow-x-auto py-1">
+        {potentialSession && (
+          <CategoryPill
+            key="foryou"
+            phEvent="foryou_navpill"
+            isActive={activeCategory === 'For You'}
+            onClick={() =>
+              handleCategoryChange(
+                'For You' as ListingCategory,
+                'foryou_navpill',
+              )
+            }
+          >
+            For You
+          </CategoryPill>
+        )}
+        <CategoryPill
+          key="all"
+          phEvent="all_navpill"
+          isActive={activeCategory === 'All'}
+          onClick={() =>
+            handleCategoryChange('All' as ListingCategory, 'all_navpill')
+          }
+        >
+          All
+        </CategoryPill>
+        {CATEGORY_NAV_ITEMS?.map((navItem) => (
+          <CategoryPill
+            key={navItem.label}
+            phEvent={navItem.pillPH}
+            isActive={activeCategory === navItem.label}
+            onClick={() =>
+              handleCategoryChange(
+                navItem.label as ListingCategory,
+                navItem.pillPH,
+              )
+            }
+          >
+            {isMd ? navItem.label : navItem.mobileLabel || navItem.label}
+          </CategoryPill>
+        ))}
+      </div>
+
+      {renderContent()}
+    </div>
+  );
+};
